Add Celery to server prototype
